requirements.txt
django-cors-headers==4.1.0
djangorestframework==3.14.0
transformers==4.33.1
python-dotenv==1.0.0
langchain==0.0.208
tiktoken==0.5.1
openai==0.28.0
torchvision
